% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellToGamma.R
\name{cellToGamma}
\alias{cellToGamma}
\title{Calculate the Goodman and Kruskal's gamma rank correlation for the specified variables on a generated random distribution.}
\usage{
cellToGamma(
  covariable = NULL,
  groups = NULL,
  labelOrder,
  indexes,
  cellCrowd,
  rank_index
)
}
\arguments{
\item{covariable}{Vector. Labels corresponding with the covariable.}

\item{groups}{Vector. Labels corresponding with the main group variable.}

\item{labelOrder}{Vector. The labels in groups in the order of the desired comparison.}

\item{indexes}{Vector. Order of the covariable as in the original data.}

\item{cellCrowd}{Vector. Number of elements to be subsampled from each group.}

\item{rank_index}{vector. The ranked values of the ordered set of groups as 1:length(labelOrder).}
}
\value{
A list of the concordant and discordant pairs calculated as per Goodman and Kruskal's gamma rank correlation.
}
\description{
`cellToGamma()` returns a list of list of the concordant and discordant value pairs calculated.
}
\details{
This function will perform a Goodman and Kruskal's gamma rank correlation, by creating first a random distribution of samples from the observed data,
in this case the distribution is generated completely at random from all the covariable labels mixed.
}
\author{
Oscar Gonzalez-Velasco
}
\keyword{internal}
