% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellToMontecarlo.R
\name{cellToMontecarlo}
\alias{cellToMontecarlo}
\title{Calculate fold-changes from a random distribution of the given data.}
\usage{
cellToMontecarlo(
  covariable = NULL,
  groups = NULL,
  labelOrder,
  indexes,
  cellCrowd
)
}
\arguments{
\item{covariable}{Vector. Labels corresponding with the covariable.}

\item{groups}{Vector. Labels corresponding with the main group variable.}

\item{labelOrder}{Vector. The labels in groups in the order of the desired comparison.}

\item{indexes}{Vector. Order of the covariable as in the original data.}

\item{cellCrowd}{Vector. Number of elements to be subsampled from each group.}
}
\value{
A list of the computated fold-changes from the random distribution created using a montecarlo test.
}
\description{
`cellToMontecarlo()` returns the fold-change from the data generated from the random sampling of the data provided.
}
\details{
This function will perform a Montecarlo simulation from the groups and covariables provided. It will compute the fold-change from this generated distribution.
}
\author{
Oscar Gonzalez-Velasco
}
\keyword{internal}
