% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Obtain the metadata}
\usage{
getMetadata(scObject = NULL, groups = NULL)
}
\arguments{
\item{scObject}{Object or data.frame containing the single-cell metadata.}

\item{groups}{Vector. Labels corresponding with the main group variable.}
}
\value{
A data.frame with the metadata from the single-cell study
}
\description{
`getMetadata()` returns a list of list of the concordant and discordant value pairs calculated.
}
\details{
This function will perform a Goodman and Kruskal's gamma rank correlation, by creating first a random distribution of samples from the observed data,
taking into account the group size, i.e. each random subset is generated using the original group distribution.
}
\author{
Oscar Gonzalez-Velasco
}
\keyword{internal}
