% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPalette.R
\name{getPalette}
\alias{getPalette}
\title{Obtain the colors of personalised palette}
\usage{
getPalette(usePalette = NULL, nColors = 20)
}
\arguments{
\item{usePalette}{Character. Name of the palette}
}
\value{
A vector of colors
}
\description{
`getMetadata()` returns a vector of nice colors.
}
\details{
This function just produces a nice color palette for the visualization functions.
}
\author{
Oscar Gonzalez-Velasco
}
\keyword{internal}
