% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAbundanceTest.R
\name{plotAbundanceTest}
\alias{plotAbundanceTest}
\title{Beautiful plot of the test results of the montecarlo simulation}
\usage{
plotAbundanceTest(tableResults = NULL, subtype_variable)
}
\arguments{
\item{covariable}{Vector. Labels corresponding with the covariable.}

\item{groups}{Vector. Labels corresponding with the main group variable.}

\item{labelOrder}{Vector. The labels in groups in the order of the desired comparison.}

\item{indexes}{Vector. Order of the covariable as in the original data.}

\item{cellCrowd}{Vector. Number of elements to be subsampled from each group.}

\item{rank_index}{vector. The ranked values of the ordered set of groups as 1:length(labelOrder).}
}
\value{
A list of the concordant and discordant pairs calculated as per Goodman and Kruskal's gamma rank correlation.
}
\description{
`plotAbundanceTest()` returns a list of list of the concordant and discordant value pairs calculated.
}
\details{
This function will perform a Goodman and Kruskal's gamma rank correlation, by creating first a random distribution of samples from the observed data,
taking into account the group size, i.e. each random subset is generated using the original group distribution.
}
\author{
Oscar Gonzalez-Velasco
}
