% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polar_chart.R
\name{polar_chart}
\alias{polar_chart}
\title{Create a polar plot from single-cell metadata.}
\usage{
polar_chart(
  scObject = NULL,
  main_variable = NULL,
  subtype_variable = NULL,
  sample_id = NULL,
  subtype_only = NULL,
  colors = NULL
)
}
\arguments{
\item{scObject}{Object or DataFrame. An object of class Single Cell Experiments or Seurat, or a dataframe containing the metadata information.}

\item{main_variable}{Character. Name of the column on the metadata dataframe containing the main variable to be used for splitting the dataset into the different bar groups (e.g.: disease_status)}

\item{subtype_variable}{Character. Name of the column on the metadata dataframe containing the covariable of interest that we want to visualize as percentages (e.g.: cell_type, time_point, ...)}

\item{sample_id}{Character. Column name containing the sample/patient id variable. If provided for tests, sampling will be done simulating the proportion variability per sample, for plots each individual will be shown.}

\item{subtype_only}{Character. Visualize only a specific class from subtype_variable. Useful if for example you only want to show the proportions of a specific cell type or subclass.}

\item{colors}{Character vector. Vector of colors defined by the user to be used as palette. If more colors than specified are required, colorRampPalette will be used to create additional colors. If not specified the LotOfCells default color palette is used.}
}
\value{
The function returns a ggplot object with the polar plot representing the population frequencies on the requested variables.
}
\description{
`polar_chart()` returns a ggplot polar plot to visualize percentages of a set of classes defined by the user.
}
\details{
This function produces a polar ggplot object with the data defined by the user.
}
\examples{
# We construct a metadata dataframe
meta.data <- data.frame(condition, cell_type, sample)
polar_chart(meta.data, main_variable = "condition", subtype_variable = "cell_type")
polar_chart(meta.data, main_variable = "condition", subtype_variable = "cell_type", sample_id = "sample")
polar_chart(meta.data, main_variable = "condition", subtype_variable = "cell_type", subtype_only = "CellType_E")
}
\author{
Oscar Gonzalez-Velasco
}
