% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffleChart.R
\name{waffle_chart}
\alias{waffle_chart}
\title{Create Waffle plots from single-cell metadata.}
\usage{
waffle_chart(
  scObject = NULL,
  main_variable = NULL,
  subtype_variable = NULL,
  sample_id = NULL,
  subtype_only = NULL
)
}
\arguments{
\item{scObject}{Object or DataFrame. An object of class Single Cell Experiments or Seurat, or a dataframe containing the metadata information.}

\item{main_variable}{Character. Name of the column on the metadata dataframe containing the main variable to be used for splitting the dataset into separated waffle plots (e.g.: disease_status)}

\item{subtype_variable}{Character. Name of the column on the metadata dataframe containing the covariable of interest that we want to visualize as percentages (e.g.: cell_type, time_point, ...)}

\item{sample_id}{Character. Column name containing the sample/patient id variable. If provided for tests, sampling will be done simulating the proportion variability per sample, for plots each individual will be shown.}

\item{subtype_only}{Character. Visualize only a specific class from subtype_variable. Useful if for example you only want to show the proportions of a specific cell type or subclass.}
}
\value{
The function returns a grid arrange of ggplots with the waffle plots representing the population frequencies on the requested variables.
}
\description{
`waffle_chart()` returns an arrange of ggplots of waffle plots to visualize percentages of a set of classes defined by the user.
}
\details{
This function produces a series of waffle plots with the data defined by the user. Each square/tile represents a 1% of the population.
}
\examples{
# We construct a metadata dataframe
meta.data <- data.frame(condition, cell_type, sample)
waffle_chart(meta.data, main_variable = "condition", subtype_variable = "cell_type")
waffle_chart(meta.data, main_variable = "condition", subtype_variable = "cell_type", sample_id = "sample")
waffle_chart(meta.data, main_variable = "condition", subtype_variable = "cell_type", subtype_only = "CellType_E")
}
\author{
Oscar Gonzalez-Velasco
}
